move supported networks into constants
