+++ packages/shared/src/lib/core/layer-2/constants/base-token-contract-address.constant.ts
@@ -1,5 +1,5 @@
 import { NetworkId } from '@core/network'
-import { SupportedEvmNetworkId } from '@core/network/enums'
+import { SupportedEvmNetworkId } from '@core/network/constants'
 
 export const BASE_TOKEN_CONTRACT_ADDRESS: Readonly<{ [id in NetworkId]?: string }> = {
     [SupportedEvmNetworkId.ShimmerEvm]: '0x1074010000000000000000000000000000000000',
