+++ packages/shared/src/lib/core/network/constants/default-coin-type.constant.ts
@@ -1,4 +1,4 @@
-import { SupportedNetworkId } from '../enums'
+import { SupportedNetworkId } from '../constants'
 import { NetworkId } from '../types'
 
 export const IOTA_COIN_TYPE = 4218
