+++ packages/shared/src/lib/core/network/constants/default-node-urls.constant.ts
@@ -1,4 +1,4 @@
-import { SupportedNetworkId } from '../enums'
+import { SupportedNetworkId } from '../constants'
 import { StardustNetworkId } from '../types'
 
 export const DEFAULT_NODE_URLS: Readonly<{ [id in StardustNetworkId]?: string[] }> = {
