+++ packages/shared/src/lib/core/network/constants/explorer-endpoints.constant.ts
@@ -1,5 +1,6 @@
-import { NetworkId } from '..'
-import { ExplorerEndpoint, SupportedNetworkId } from '../enums'
+import { ExplorerEndpoint } from '../enums'
+import { NetworkId } from '../types'
+import { SupportedNetworkId } from './supported-network-id.constant'
 
 export const EXPLORER_ENDPOINTS: Readonly<{ [key in NetworkId]?: { [key in ExplorerEndpoint]?: string } }> = {
     [SupportedNetworkId.Iota]: {
