+++ packages/shared/src/lib/core/network/constants/index.ts
@@ -18,3 +18,4 @@ export * from './max-network-name-length.constant'
 export * from './milestone-not-found.constant'
 export * from './network-status-poll-interval.constant'
 export * from './seconds-per-milestone.constant'
+export * from './supported-network-id.constant'
