+++ packages/shared/src/lib/core/network/enums/index.ts
@@ -4,6 +4,5 @@ export * from './explorer-api-endpoint.enum'
 export * from './explorer-endpoint.enum'
 export * from './network-health.enum'
 export * from './network-namespace.enum'
-export * from './supported-network-id.enum'
 export * from './stardust-network-name.enum'
 export * from './token-supply.enum'
