+++ packages/shared/src/lib/core/network/tests/network.test.ts
@@ -1,5 +1,5 @@
 import { INode } from '@iota/sdk/out/types'
-import { SupportedNetworkId } from '../enums'
+import { SupportedNetworkId } from '../constants'
 import { NetworkId } from '../types'
 import { getDefaultNodes, isSupportedNetworkId } from '../utils'
