+++ packages/shared/src/lib/core/network/utils/isSupportedNetworkId.ts
@@ -1,4 +1,4 @@
-import { SupportedNetworkId } from '../enums'
+import { SupportedNetworkId } from '../constants'
 import { NetworkId } from '../types'
 
 export function isSupportedNetworkId(networkId: NetworkId): boolean {
