+++ packages/shared/src/lib/core/wallet/tests/getOutputParameters.test.ts
@@ -1,6 +1,5 @@
-import { SupportedNetworkId } from '@core/network/enums'
 import { FALLBACK_ESTIMATED_GAS } from '@core/layer-2/constants'
-import { DEFAULT_EVM_NETWORK_CONFIGURATIONS } from '@core/network/constants'
+import { DEFAULT_EVM_NETWORK_CONFIGURATIONS, SupportedNetworkId } from '@core/network/constants'
 import { getOutputParameters } from '../utils'
 import { ReturnStrategy, SubjectType } from '../enums'
 import { IToken, IPersistedToken } from '@core/token/interfaces'
